% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc_load.R
\name{atc_load}
\alias{atc_load}
\title{Automatically load my most used packages}
\usage{
atc_load()
}
\description{
Add more packages to the body to facilitate loading of
multiple packages at the beginning of an R session
}
