% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atc_mw_power.R
\name{atc_mw_power}
\alias{atc_mw_power}
\title{Power Analysis through resampling and Monte Carlo style simulation for Mann-Whitney tests}
\usage{
atc_mw_power(control_data, experiment_data, reps = 1000)
}
\arguments{
\item{control_data}{A vector of the control data to be resampled}

\item{experiment_data}{A vector of the experiment data to be resampled}

\item{reps}{The number of iteration for the bootstrap, default = 1000}
}
\description{
Calculate power as the percentage of tests yielding the same results through bootstraps
}
