% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Analog of Matlab function to clear all variables from the global environment}
\usage{
clear()
}
\description{
Analog of Matlab function to clear all variables from the global environment
}
