% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearvars_except.R
\name{clearvars_except}
\alias{clearvars_except}
\title{Analog of Matlab function to clear all variables from the global environment}
\usage{
clearvars_except(...)
}
\description{
Analog of Matlab function to clear all variables from the global environment
}
