% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_summary.R
\name{se_summary}
\alias{se_summary}
\title{Dplyr-friendly function to calculate Mean, SD, and SEM through summarize}
\usage{
se_summary(.data, .var)
}
\description{
Dplyr-friendly function to calculate Mean, SD, and SEM through summarize
}
